Extract CLI from cephlcmlib to cephlcm_cli package
