+++ cephlcmcli/cephlcm_cli/cli/decorators.py
@@ -9,10 +9,10 @@ import os
 
 import click
 import six
-
 from cephlcmlib import exceptions
-from cephlcmlib.cli import utils
-from cephlcmlib.cli import param_types
+
+from cephlcm_cli.cli import utils
+from cephlcm_cli.cli import param_types
 
 
 def catch_errors(func):
