+++ cephlcmcli/cephlcm_cli/cli/main.py
@@ -6,9 +6,9 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 
 import click
-
 import cephlcmlib
-from cephlcmlib.cli import utils
+
+from cephlcm_cli.cli import utils
 
 
 CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
@@ -103,11 +103,11 @@ def cli_group(func):
     return func
 
 
-from cephlcmlib.cli import cluster  # NOQA
-from cephlcmlib.cli import execution  # NOQA
-from cephlcmlib.cli import permission  # NOQA
-from cephlcmlib.cli import playbook  # NOQA
-from cephlcmlib.cli import playbook_configuration  # NOQA
-from cephlcmlib.cli import role  # NOQA
-from cephlcmlib.cli import server  # NOQA
-from cephlcmlib.cli import user  # NOQA
+from cephclm_cli.cli import cluster  # NOQA
+from cephclm_cli.cli import execution  # NOQA
+from cephclm_cli.cli import permission  # NOQA
+from cephclm_cli.cli import playbook  # NOQA
+from cephclm_cli.cli import playbook_configuration  # NOQA
+from cephclm_cli.cli import role  # NOQA
+from cephclm_cli.cli import server  # NOQA
+from cephclm_cli.cli import user  # NOQA
