+++ cephlcmcli/cephlcm_cli/cli/param_types.py
@@ -7,7 +7,7 @@ from __future__ import unicode_literals
 
 import click.types
 
-from cephlcmlib.cli import utils
+from cephlcm_cli.cli import utils
 
 
 class CSVParamType(click.types.ParamType):
