+++ cephlcmcli/cephlcm_cli/cli/role.py
@@ -7,10 +7,10 @@ from __future__ import unicode_literals
 
 import click
 
-from cephlcmlib.cli import cli_group
-from cephlcmlib.cli import decorators
-from cephlcmlib.cli import param_types
-from cephlcmlib.cli import utils
+from cephlcm_cli.cli import main
+from cephlcm_cli.cli import decorators
+from cephlcm_cli.cli import param_types
+from cephlcm_cli.cli import utils
 
 
 def permissions_to_dict(permissions):
@@ -24,7 +24,7 @@ def permissions_to_list(permissions):
     ]
 
 
-@cli_group
+@main.cli_group
 def role():
     """Role subcommands."""
