+++ cephlcmcli/cephlcm_cli/cloud_config.py
