+++ cephlcmcli/setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
-name = cephlcmlib
+name = cephlcm-cli
 version = 0.1
-summary = Ceph lifecycle management client library
+summary = Ceph lifecycle management CLI
 maintainer = Mirantis
 mainteiner-email = sarkhipov@mirantis.com
 author = Mirantis
@@ -25,6 +25,11 @@ classifier =
 [files]
 packages =
   cephlcmlib
+
+[entry_points]
+console_scripts =
+  cephlcm = cephlcm_cli.cli:cli
+  cephlcm-cloud-config = cephlcm_cli.cloud_config:main
 
 [global]
 setup-hooks =
